pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "motion-mirror"
version = "0.2a0"
description = "Local-first motion transfer pipeline for animating a character image from a reference video."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "gradio>=5,<6",
  "typer>=0.12,<1",
  "opencv-python>=4.9",
  "numpy>=1.26",
  "rembg>=2.0",
  "static-ffmpeg>=2.5",
  "ffmpeg-python>=0.2",
  "huggingface-hub>=0.23",
  "rich>=13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]
cuda = [
  "torch>=2.2",
  "torchvision>=0.17",
  "onnxruntime-gpu>=1.18",
  "rtmlib>=0.0.13",
]
cpu = [
  "torch>=2.2",
  "torchvision>=0.17",
  "onnxruntime>=1.18",
  "rtmlib>=0.0.13",
]
# v0.2a additions
# Note: lightx2v must be installed from GitHub: pip install git+https://github.com/ModelTC/LightX2V.git
# Note: sam2 must be installed from GitHub: pip install git+https://github.com/facebookresearch/sam2.git
# RAFT weights are auto-downloaded by torchvision — no extra package needed.
gpu-inference = [
  "diffusers>=0.33",
  "transformers",
  "accelerate",
  "ftfy",
  "Pillow",
]
lightx2v = [
  "lightx2v @ git+https://github.com/ModelTC/LightX2V.git",
  "pyzmq>=26",
]

[project.scripts]
motion-mirror = "motion_mirror.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"motion_mirror" = ["presets/*.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "gpu: marks tests requiring a CUDA GPU (deselect with -m 'not gpu')",
  "slow: marks tests that take more than 30 seconds",
]
